+++ Src/AssemblyInfo.cs
@@ -4,7 +4,7 @@ using System.Runtime.InteropServices;
 using System;
 
 [assembly: AssemblyProduct("Mdbc")]
-[assembly: AssemblyVersion("2.1.3")]
+[assembly: AssemblyVersion("3.0.0")]
 [assembly: AssemblyTitle("Mdbc module - MongoDB Cmdlets for PowerShell")]
 [assembly: AssemblyCompany("https://github.com/nightroman/Mdbc")]
 [assembly: AssemblyCopyright("Copyright (c) 2011-2012 Roman Kuzmin")]